all the sensors

+++ CarServer/GlobalV.c
@@ -1,3 +1,3 @@
-double car_speed = 0;
-double change_speed =0;
+int car_speed = 0;
+int change_speed =0;
 double distance_to_crash = 0;
